translation-service/requirements.txt
flask
flask_socketio
flask_cors
flask_sqlalchemy
python_dotenv
eventlet
deep-translator